+++ mne/mixed_norm/__init__.py
@@ -1,0 +1,7 @@
+"""Non-Linear inverse solvers based on Mixed Norm Estimates (MxNE)"""
+
+# Author: Alexandre Gramfort <gramfort@nmr.mgh.harvard.edu>
+#
+# License: Simplified BSD
+
+from .inverse import mixed_norm